fix: fix reload crashes.
Former-commit-id: 73e901e13fb62dd4cbddadff882d12533f33b8a8 [formerly 7ceb0841d3f1c493c6a147d9eb0e14f618cfaa42]
Former-commit-id: 9121f53a31abc51be35946f456f053a0e7d3616d

+++ bridge/foundation/ui_command_queue.cc
@@ -7,7 +7,7 @@
 
 namespace foundation {
 
-void UICommandTaskMessageQueue::registerCommand(int64_t id, int8_t type, NativeString **args, size_t length, void* nativePtr) {
+void UICommandTaskMessageQueue::registerCommand(int64_t id, int32_t type, NativeString **args, size_t length, void* nativePtr) {
   auto item = new UICommandItem(id, type, args, length, nativePtr);
   queue.emplace_back(item);
 }
